[CBD] Extract WebBundle logic from Archive into WebBundleHelper
Move all WebBundle-specific code (detection, parsing, extraction,
listing) out of Archive.cs into a dedicated WebBundleHelper class.
Archive now delegates to WebBundleHelper for web bundle operations.
